requirements.txt
# ── FastAPI Backend (Render) ──────────────────────────
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.44.0
joblib>=1.3.0
psycopg2-binary>=2.9.0

# ── Streamlit Frontend (Streamlit Community Cloud) ────
streamlit>=1.32.0
streamlit-authenticator>=0.3.3
plotly>=5.18.0
requests>=2.31.0
bcrypt>=4.0.0
